almost working before I remove array support
