+++ test/ecto/integration/crud_test.exs
@@ -219,7 +219,7 @@ defmodule Ecto.Integration.CrudTest do
     test "can handle in" do
       TestRepo.insert!(%Account{name: "hi"})
       assert [] = TestRepo.all(from(a in Account, where: a.name in ["404"]))
-      assert [_] = TestRepo.all(from(a in Account, where: a.name in ["hi"]))
+      assert [_] = TestRepo.all(from(a in Account, where: a.name in [^"hi"]))
     end
 
     test "handles case sensitive text" do
